+++ mu4/importexport/internal/musicxml/exportxml.cpp
@@ -1791,7 +1791,7 @@ static void fermata(const Fermata* const a, XmlWriter& xml)
              || id == SymId::fermataVeryLongAbove || id == SymId::fermataVeryLongBelow) {
         xml.tag(tagName, "square");
     } else {
-        qDebug("unknown fermata sim id %d", id);
+        qDebug("unknown fermata sim id %d", static_cast<int>(id));
     }
 }
 
@@ -2893,7 +2893,7 @@ void ExportMusicXml::chordAttributes(Chord* chord, Notations& notations, Technic
             && symIdToOrnam(sid) == ""
             && symIdToTechn(sid) == ""
             && !isLaissezVibrer(sid)) {
-            qDebug("unknown chord attribute %d %s", sid, qPrintable(a->userName()));
+            qDebug("unknown chord attribute %d %s", static_cast<int>(sid), qPrintable(a->userName()));
         }
     }
 }
